# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 833 Tidyups Booking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff members
      parameters:
        - in: query
          name: activeOnly
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: List of staff members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Staff"

    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a new staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Staff member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/me:
    get:
      operationId: getStaffMe
      tags: [staff]
      summary: Get the staff record linked to the authenticated Clerk account
      responses:
        "200":
          description: Staff record for the current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
        "404":
          description: No staff record is linked to this Clerk account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/{id}:
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update a staff member
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdate"
      responses:
        "200":
          description: Staff member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/{id}/schedule:
    get:
      operationId: getStaffSchedule
      tags: [staff]
      summary: Get a staff member's bookings for a given date
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
        - in: query
          name: date
          required: true
          schema:
            type: string
            description: Date in YYYY-MM-DD format
      responses:
        "200":
          description: Staff schedule for the date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/{id}/location:
    post:
      operationId: postStaffLocation
      tags: [staff]
      summary: Post cleaner's live GPS location
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lat, lng]
              properties:
                lat:
                  type: number
                lng:
                  type: number
                accuracy:
                  type: number
      responses:
        "200":
          description: Location recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /schedule:
    get:
      operationId: getDaySchedule
      tags: [staff]
      summary: Get all staff schedules for a given date
      parameters:
        - in: query
          name: date
          required: true
          schema:
            type: string
            description: Date in YYYY-MM-DD format
      responses:
        "200":
          description: All staff schedules for the date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffDaySchedule"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Booking:
      type: object
      required:
        - id
        - firstName
        - lastName
        - phone
        - address
        - city
        - serviceType
        - bedrooms
        - bathrooms
        - scheduledDate
        - scheduledTime
        - frequency
        - status
        - createdAt
      properties:
        id:
          type: number
          format: int32
        staffId:
          type: ["number", "null"]
          format: int32
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        address:
          type: string
        city:
          type: string
        province:
          type: string
        postalCode:
          type: ["string", "null"]
        serviceType:
          type: string
          enum: [standard_clean, deep_clean, move_in_out, post_construction]
        bedrooms:
          type: number
          format: int32
        bathrooms:
          type: number
          format: int32
        extras:
          type: array
          items:
            type: string
        scheduledDate:
          type: string
          description: Date in YYYY-MM-DD format
        scheduledTime:
          type: string
          description: Time in HH:MM format
        frequency:
          type: string
          enum: [one_time, weekly, biweekly, monthly]
        estimatedPrice:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled]
        jobberJobId:
          type: ["string", "null"]
        jobberSyncStatus:
          type: string
          enum: [not_started, pending, synced, failed]
          description: >
            Jobber sync state. not_started = never attempted, pending = in
            progress, synced = succeeded, failed = error (see jobberSyncError).
        jobberSyncError:
          type: ["string", "null"]
          description: Error message from the last failed Jobber sync attempt.
        createdAt:
          type: string
          format: date-time
        hasTranscript:
          type: boolean
          description: True when at least one call transcript is attached to this booking.
...
    Staff:
      type: object
      required:
        - id
        - name
        - role
        - active
        - createdAt
      properties:
        id:
          type: number
          format: int32
        name:
          type: string
        role:
          type: string
          enum: [cleaner, lead_cleaner, supervisor]
        phone:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    StaffInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
          enum: [cleaner, lead_cleaner, supervisor]
          default: cleaner
        phone:
          type: string
        active:
          type: boolean
          default: true
    StaffUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [cleaner, lead_cleaner, supervisor]
        phone:
          type: string
        active:
          type: boolean
        clerkUserId:
          type: string
          nullable: true
          description: >
            Clerk user ID to link to this staff record. Dispatcher-only field.
            Set to null to unlink. When set, the staff member can sign in to the
            cleaner app and their schedule will load automatically.
        homeAddress:
          type: string
          nullable: true
          description: Street address for the cleaner's home base (used for map proximity).
        homeLat:
          type: number
          nullable: true
          description: Latitude of the home address.
        homeLng:
          type: number
          nullable: true
          description: Longitude of the home address.
    StaffDaySchedule:
      type: object
      required:
        - staff
        - bookings
      properties:
        staff:
          $ref: "#/components/schemas/Staff"
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"