# lib/api-zod/src/generated/types/staffUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * 833 Tidyups Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { StaffUpdateRole } from './staffUpdateRole';

export interface StaffUpdate {
  name?: string;
  role?: StaffUpdateRole;
  phone?: string;
  active?: boolean;
  /** Clerk user ID to link to this staff record. Dispatcher-only field. Set to null to unlink. When set, the staff member can sign in to the cleaner app and their schedule will load automatically. */
  clerkUserId?: string | null;
  /** Street address for the cleaner's home base (used for map proximity). */
  homeAddress?: string | null;
  /** Latitude of the home address. */
  homeLat?: number | null;
  /** Longitude of the home address. */
  homeLng?: number | null;
}